abstract Pools/PoolMemberships
